Update generate_counterexample_typeslines.py
Added the ability to change the ranges of what is considered a "good pose" and a "bad pose". Before the script only labeled good poses as <2 RMSD, and all other poses as bad.

Now you can set the good pose threshold with --good_pose_thresh, and set the bad pose threshold with --bad_pose_thresh. 

This means poses <good_pose_thresh are labelled with 1, poses >=bad_pose_thresh are labelled with 0, and poses with RMSD between good_pose_thresh and bad_pose_thresh are discarded.